changed import of libraries so that the jar can be build not only locally
