+++ approaches/optimistic/src/main/java/de/hpi/cluster/actors/Master.java
@@ -14,6 +14,9 @@ import de.hpi.utils.perfromance.PerformanceTracker;
 import de.hpi.evaluation.ConsoleOutputEvaluator;
 import de.hpi.evaluation.GoldStandardEvaluator;
 import de.hpi.partitioning.Md5HashRouter;
+import de.hpi.rdse.der.evaluation.ConsoleOutputEvaluator;
+import de.hpi.rdse.der.evaluation.GoldStandardEvaluator;
+import de.hpi.rdse.der.partitioning.Md5HashRouter;
 import de.hpi.utils.data.CSVService;
 import de.hpi.utils.perfromance.PerformanceTracker;
 import lombok.AllArgsConstructor;
