+++ approaches/optimistic/src/main/java/de/hpi/cluster/actors/Worker.java
@@ -13,14 +13,14 @@ import akka.event.LoggingAdapter;
 import de.hpi.cluster.ClusterMaster;
 import de.hpi.cluster.messages.InfoObject;
 import de.hpi.cluster.messages.interfaces.Blocking;
-import de.hpi.der.dude.DuplicateDetector;
-import de.hpi.der.dude.SimpleDuplicateDetector;
-import de.hpi.der.similarity.UniversalComparator;
-import de.hpi.der.similarity.numeric.AbsComparator;
-import de.hpi.der.similarity.numeric.NumberComparator;
-import de.hpi.der.similarity.string.JaroWinklerComparator;
-import de.hpi.der.similarity.string.StringComparator;
-import de.hpi.partitioning.Md5HashRouter;
+import de.hpi.rdse.der.dude.DuplicateDetector;
+import de.hpi.rdse.der.dude.SimpleDuplicateDetector;
+import de.hpi.rdse.der.partitioning.Md5HashRouter;
+import de.hpi.rdse.der.similarity.UniversalComparator;
+import de.hpi.rdse.der.similarity.numeric.AbsComparator;
+import de.hpi.rdse.der.similarity.numeric.NumberComparator;
+import de.hpi.rdse.der.similarity.string.JaroWinklerComparator;
+import de.hpi.rdse.der.similarity.string.StringComparator;
 import lombok.AllArgsConstructor;
 import lombok.Data;
