--- approaches/shared/duplicate-detection/src/main/java/de/hpi/der/dude/AbstractDuplicateDetector.java
@@ -1,15 +1,0 @@
-package de.hpi.der.dude;
-
-import de.hpi.der.similarity.UniversalComparator;
-
-public abstract class AbstractDuplicateDetector implements DuplicateDetector {
-
-    UniversalComparator comparator;
-    double similarityThreshold;
-
-    public AbstractDuplicateDetector(UniversalComparator comparator, double similarityThreshold) {
-        this.comparator = comparator;
-        this.similarityThreshold = similarityThreshold;
-    }
-
-}
