--- approaches/shared/duplicate-detection/src/main/java/de/hpi/der/similarity/numeric/NumberComparator.java
@@ -1,12 +1,0 @@
-package de.hpi.der.similarity.numeric;
-
-public interface NumberComparator {
-
-    /**
-     * Compares two given doubles
-     * @param d1 first double
-     * @param d2 second double
-     * @return A double between 0 and 1 indicating how similar the two numbers are
-     */
-    double compare(double d1, double d2);
-}
