--- approaches/shared/duplicate-detection/src/main/java/de/hpi/der/similarity/string/NGramComparator.java
@@ -1,12 +1,0 @@
-package de.hpi.der.similarity.string;
-
-import info.debatty.java.stringsimilarity.NGram;
-
-public class NGramComparator implements StringComparator {
-
-    private NGram ng = new NGram();
-
-    public double compare(String s1, String s2) {
-        return 1 - ng.distance(s1, s2);
-    }
-}
