+++ approaches/shared/duplicate-detection/src/main/java/de/hpi/rdse/der/dude/AbstractDuplicateDetector.java
@@ -1,0 +1,15 @@
+package de.hpi.rdse.der.dude;
+
+import de.hpi.rdse.der.similarity.UniversalComparator;
+
+public abstract class AbstractDuplicateDetector implements DuplicateDetector {
+
+    UniversalComparator comparator;
+    double similarityThreshold;
+
+    public AbstractDuplicateDetector(UniversalComparator comparator, double similarityThreshold) {
+        this.comparator = comparator;
+        this.similarityThreshold = similarityThreshold;
+    }
+
+}
