+++ approaches/shared/duplicate-detection/src/main/java/de/hpi/rdse/der/similarity/string/JaroWinklerComparator.java
@@ -1,0 +1,12 @@
+package de.hpi.rdse.der.similarity.string;
+
+import info.debatty.java.stringsimilarity.JaroWinkler;
+
+public class JaroWinklerComparator implements StringComparator {
+
+    private JaroWinkler jw = new JaroWinkler();
+
+    public double compare(String s1, String s2) {
+        return jw.similarity(s1, s2);
+    }
+}
