+++ approaches/shared/duplicate-detection/src/main/java/de/hpi/rdse/der/similarity/string/NGramComparator.java
@@ -1,0 +1,12 @@
+package de.hpi.rdse.der.similarity.string;
+
+import info.debatty.java.stringsimilarity.NGram;
+
+public class NGramComparator implements StringComparator {
+
+    private NGram ng = new NGram();
+
+    public double compare(String s1, String s2) {
+        return 1 - ng.distance(s1, s2);
+    }
+}
