--- approaches/shared/duplicate-detection/src/test/java/de/hpi/der/combinations/CombinationsTest.java
@@ -1,6 +1,0 @@
-package de.hpi.der.combinations;
-
-public class CombinationsTest {
-
-    // TODO write Test for this code
-}
