--- approaches/shared/duplicate-detection/src/test/java/de/hpi/der/dude/SimpleDuplicateDetectorTest.java
@@ -1,6 +1,0 @@
-package de.hpi.der.dude;
-
-public class SimpleDuplicateDetectorTest {
-
-    // TODO
-}
