+++ approaches/shared/duplicate-detection/src/test/java/de/hpi/rdse/der/combinations/CombinationsTest.java
@@ -1,0 +1,6 @@
+package de.hpi.rdse.der.combinations;
+
+public class CombinationsTest {
+
+    // TODO write Test for this code
+}
