+++ approaches/shared/duplicate-detection/src/test/java/de/hpi/rdse/der/dude/SimpleDuplicateDetectorTest.java
@@ -1,0 +1,6 @@
+package de.hpi.rdse.der.dude;
+
+public class SimpleDuplicateDetectorTest {
+
+    // TODO
+}
