+++ approaches/shared/evaluation/pom.xml
@@ -8,11 +8,16 @@
     <artifactId>evaluation</artifactId>
     <version>1.0</version>
 
+    <properties>
+        <project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
+    </properties>
+
     <build>
         <plugins>
             <plugin>
                 <groupId>org.apache.maven.plugins</groupId>
                 <artifactId>maven-compiler-plugin</artifactId>
+                <version>2.3.2</version>
                 <configuration>
                     <source>1.8</source>
                     <target>1.8</target>
@@ -48,11 +53,9 @@
             <scope>compile</scope>
         </dependency>
         <dependency>
-            <groupId>der.rdse.hpi.de</groupId>
+            <groupId>de.hpi.rdse.der</groupId>
             <artifactId>setoperations</artifactId>
             <version>1.0</version>
-            <scope>system</scope>
-            <systemPath>${basedir}/../set-operations/target/setoperations-1.0.jar</systemPath>
         </dependency>
     </dependencies>
