+++ docker/testing/data/data_gold.csv
@@ -1,0 +1,113 @@
+"class","id_1","id_2"
+'0', 1, 2
+'1', 3, 4
+'2', 5, 6
+'3', 7, 8
+'4', 9, 10
+'5', 11, 12
+'6', 13, 14
+'7', 15, 16
+'8', 17, 18
+'9', 19, 20
+'10', 21, 22
+'11', 23, 24
+'12', 25, 26
+'13', 27, 28
+'14', 29, 30
+'15', 31, 32
+'16', 33, 34
+'17', 35, 36
+'18', 37, 38
+'19', 39, 40
+'20', 41, 42
+'21', 43, 44
+'22', 45, 46
+'23', 47, 48
+'24', 49, 50
+'25', 51, 52
+'26', 53, 54
+'27', 55, 56
+'28', 57, 58
+'29', 59, 60
+'30', 61, 62
+'31', 63, 64
+'32', 65, 66
+'33', 67, 68
+'34', 69, 70
+'35', 71, 72
+'36', 73, 74
+'37', 75, 76
+'38', 77, 78
+'39', 79, 80
+'40', 81, 82
+'41', 83, 84
+'42', 85, 86
+'43', 87, 88
+'44', 89, 90
+'45', 91, 92
+'46', 93, 94
+'47', 95, 96
+'48', 97, 98
+'49', 99, 100
+'50', 101, 102
+'51', 103, 104
+'52', 105, 106
+'53', 107, 108
+'54', 109, 110
+'55', 111, 112
+'56', 113, 114
+'57', 115, 116
+'58', 117, 118
+'59', 119, 120
+'60', 121, 122
+'61', 123, 124
+'62', 125, 126
+'63', 127, 128
+'64', 129, 130
+'65', 131, 132
+'66', 133, 134
+'67', 135, 136
+'68', 137, 138
+'69', 139, 140
+'70', 141, 142
+'71', 143, 144
+'72', 145, 146
+'73', 147, 148
+'74', 149, 150
+'75', 151, 152
+'76', 153, 154
+'77', 155, 156
+'78', 157, 158
+'79', 159, 160
+'80', 161, 162
+'81', 163, 164
+'82', 165, 166
+'83', 167, 168
+'84', 169, 170
+'85', 171, 172
+'86', 173, 174
+'87', 175, 176
+'88', 177, 178
+'89', 179, 180
+'90', 181, 182
+'91', 183, 184
+'92', 185, 186
+'93', 187, 188
+'94', 189, 190
+'95', 191, 192
+'96', 193, 194
+'97', 195, 196
+'98', 197, 198
+'99', 199, 200
+'100', 201, 202
+'101', 203, 204
+'102', 205, 206
+'103', 207, 208
+'104', 209, 210
+'105', 211, 212
+'106', 213, 214
+'107', 215, 216
+'108', 217, 218
+'109', 219, 220
+'110', 221, 222
+'111', 223, 224
